requirements.txt
google-adk>=1.15.0
pydantic>=2.0.0
loguru>=0.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiofiles>=23.0.0
python-dotenv>=1.0.0

flask>=3.0.0
flask-socketio>=5.3.0
aiohttp>=3.9.0
